Have refactored it slightly
