+++ src/Refactored/Chihuahua.java
@@ -1,0 +1,20 @@
+package Refactored;
+
+public class Chihuahua implements DogInterface {
+	
+	public void name() {
+		System.out.println("Chispa");
+	}
+
+	public void species() {
+		System.out.println("Is a Dog.");
+	}
+
+	public void bark() {
+		System.out.println("Goes WOOF!");
+	}
+
+	public void walk() {
+		System.out.println("Has been for a W.A.L.K");
+	}
+}
