+++ src/Refactored/DogInterface.java
@@ -1,0 +1,13 @@
+package Refactored;
+
+public interface DogInterface {
+	// Interfaces assume abstraction but we can add abstract if we want
+	// But we cannot add implemented methods.
+	abstract void name();
+
+	void species();
+
+	void bark();
+
+	void walk();
+}
