+++ src/Refactored/Labrador.java
@@ -1,0 +1,21 @@
+package Refactored;
+
+
+public class Labrador implements DogInterface {
+	
+	public void name() {
+		System.out.println("Goldie");
+	}
+
+	public void species() {
+		System.out.println("Is a Dog.");
+	}
+
+	public void bark() {
+		System.out.println("Goes WOOF!");
+	}
+
+	public void walk() {
+		System.out.println("Has been for a W.A.L.K");
+	}
+}
